Implemented prettier and eslint (airbnb).
