--- src/components/App.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-import { BrowserRouter } from 'react-router-dom'
-import MainRouter from './MainRouter.jsx'
-
-const App = () => (
-  <BrowserRouter>
-    <MainRouter />
-  </BrowserRouter>
-)
-
-export default App
