+++ src/components/core/Home.js
@@ -1,0 +1,5 @@
+import React, { Fragment } from 'react';
+
+const Home = () => <Fragment>Home Component</Fragment>;
+
+export default Home;
