--- src/components/core/Home.jsx
@@ -1,5 +1,0 @@
-import React, { Component } from 'react'
-
-const Home = () => <React.Fragment>Home Component</React.Fragment>
-
-export default Home
